moved Equality & EqualityOperations to a separate class
