+++ java-backend/src/main/java/org/kframework/backend/java/kil/Bottom.java
@@ -2,6 +2,8 @@
 package org.kframework.backend.java.kil;
 
 import java.util.EnumMap;
+
+import org.kframework.backend.java.symbolic.Equality;
 import org.kframework.backend.java.symbolic.Matcher;
 import org.kframework.backend.java.symbolic.Transformer;
 import org.kframework.backend.java.symbolic.Unifier;
@@ -12,7 +14,7 @@ import com.google.common.collect.Maps;
 
 /**
  * Refers to a computation which never completes successfully.
- * A {@link org.kframework.backend.java.symbolic.SymbolicConstraint.Equality} instance between
+ * A {@link org.kframework.backend.java.symbolic.Equality} instance between
  * bottom and anything else is false and makes the entire constraint false.
  *
  * @see org.kframework.backend.java.symbolic.SymbolicConstraint
