+++ java-backend/src/main/java/org/kframework/backend/java/symbolic/PrePostVisitor.java
@@ -370,7 +370,7 @@ public class PrePostVisitor implements Visitor {
             entry.getKey().accept(this);
             entry.getValue().accept(this);
         }
-        for (SymbolicConstraint.Equality equality : node.equalities()) {
+        for (Equality equality : node.equalities()) {
             equality.leftHandSide().accept(this);
             equality.rightHandSide().accept(this);
         }
