+++ java-backend/src/test/java/org/kframework/backend/java/symbolic/UseSMTTest.java
@@ -16,7 +16,7 @@ import org.kframework.backend.java.kil.GlobalContext;
 import org.kframework.backend.java.kil.KLabelConstant;
 import org.kframework.backend.java.kil.Rule;
 import org.kframework.backend.java.kil.TermContext;
-import org.kframework.backend.java.symbolic.SymbolicConstraint.EqualityOperations;
+import org.kframework.backend.java.symbolic.Equality.EqualityOperations;
 import org.kframework.kil.loader.Context;
 import org.kframework.utils.options.SMTOptions;
 import org.kframework.utils.options.SMTSolver;
@@ -38,7 +38,7 @@ public class UseSMTTest {
     Definition definition;
 
     @Mock
-    EqualityOperations equalityOps;
+    Equality.EqualityOperations equalityOps;
 
     @Before
     public void setUp() {
